+++ app/kumactl/cmd/install/install.go
@@ -13,6 +13,7 @@ func NewInstallCmd(pctx *kumactl_cmd.RootContext) *cobra.Command {
 	}
 	// sub-commands
 	cmd.AddCommand(newInstallControlPlaneCmd(pctx))
+	cmd.AddCommand(newInstallMetrics())
 	cmd.AddCommand(newInstallDatabaseSchemaCmd())
 	return cmd
 }