+++ app/kumactl/pkg/install/k8s/metrics/templates_vfsdata_test.go
@@ -1,6 +1,6 @@
 // +build !dev
 
-package controlplane_test
+package metrics_test
 
 import (
 	"io/ioutil"
@@ -10,20 +10,20 @@ import (
 	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
 
-	controlplane "github.com/Kong/kuma/app/kumactl/pkg/install/k8s/control-plane"
+	"github.com/Kong/kuma/app/kumactl/pkg/install/k8s/metrics"
 	"github.com/Kong/kuma/pkg/test/vfsgen"
 )
 
 var _ = Describe("Templates", func() {
 
 	kumactlSrcDir := filepath.Join("..", "..", "..", "..")
-	controlplaneTemplatesDir := controlplane.TemplatesDir(kumactlSrcDir)
-	controlplaneTemplatesTestEntries := vfsgen.GenerateEntries(controlplaneTemplatesDir)
+	metricsTemplatesDir := metrics.TemplatesDir(kumactlSrcDir)
+	metricsTemplatesTestEntries := vfsgen.GenerateEntries(metricsTemplatesDir)
 
 	DescribeTable("generated Go code must be in sync with the original template files",
 		func(given vfsgen.FileTestCase) {
 			// when
-			file, err := controlplane.Templates.Open(given.Filename)
+			file, err := metrics.Templates.Open(given.Filename)
 			// then
 			Expect(err).ToNot(HaveOccurred())
 
@@ -32,8 +32,8 @@ var _ = Describe("Templates", func() {
 			// then
 			Expect(err).ToNot(HaveOccurred())
 			// and
-			Expect(string(actualContents)).To(Equal(string(given.ExpectedContents)), "generated Go code is no longer in sync with the original template files. To re-generate it, run `make generate/kumactl/install/k8s/control-plane`")
+			Expect(string(actualContents)).To(Equal(string(given.ExpectedContents)), "generated Go code is no longer in sync with the original template files. To re-generate it, run `make generate/kumactl/install/k8s/metrics`")
 		},
-		controlplaneTemplatesTestEntries...,
+		metricsTemplatesTestEntries...,
 	)
 })