+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Changes:
   [#607](https://github.com/Kong/kuma/pull/607)
 * fix: remove the requirement to a `Pod` to explicitly list container ports in a case where a `Service` defines target port by number
   [#605](https://github.com/Kong/kuma/pull/605)
+* feature: kumactl install metrics for one line Prometheus and Grafana install on K8S
+  [#604](https://github.com/Kong/kuma/pull/604)
 * feature: order of meta in REST Resource JSON 
   [#600](https://github.com/Kong/kuma/pull/600)
 * feature: extend embedded gRPC Access Log Server to support the entire Envoy access log format